requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyodbc
alembic
passlib[bcrypt]
python-jose
pytest
